build: switch over to nob
i was getting tired of fighting with bash over regex.
although, i had to sacrifice platform-specific code compilation, but
it's kind of becoming deprecated on its own already and is subject to
removal in near future, as crossplatform versions work on both platforms
without issues, regardless of whether C or C++ is used. the only real
issue are permissions and that one inconsistency mentioned on the README